Build: Add build analytics option for research
This is a script that sends hardware info and build times to a server in order to aggregate it, and be able to make a wizard to tell $
This is ONLY info about how long it took to build and the hardware related to build performance
It is defaulted off and the way to turn it on is exporting SHARE_BUILD_ANALYTICS to true in a device tree.

The info collected is as follows:
The CPU Model
The number of physical CPU's installed
The distro you are using
Whether you are using ccache, and the size of it
The SSD Names in your system (e.g. /dev/sda)
The HDD Names in you system (e.g. /dev/sdb)
The disk output is going to
The disk input is coming from
The total memory installed
The platform(as shown when you run lunch $device)
Whether you use a prebuilt chromium
And the build time in seconds

You can see the source in this commit, and also see the server-side script on
https://github.com/USA-RedDragon/BuildAnalytics

I hope you will export this to true in order to further this research and give a good, accurate representations of Android build time